feat: redisson 분산락 적용 코드 작성

+++ spring-redisson/src/main/java/com/cooper/springredission/dto/PromotionCreateRequest.java
@@ -10,6 +10,6 @@ import lombok.NoArgsConstructor;
 public class PromotionCreateRequest {
 
     private String promotionName;
-    private Long ticketMaxLimit;
+    private int ticketMaxLimit;
 
 }
